+++ packages/bio-parsers/CHANGELOG.md
@@ -1,8 +1,20 @@
 # Changelog
 
-## 0.4.34 (2025-01-22)
+## 0.4.37 (2025-02-28)
 
 - fix: sync the description change into genbank 'DESCRIPTION'
+
+## 0.4.36 (2026-02-25)
+
+- fix: format long note as multiple lines in Genbank string
+
+## 0.4.35 (2026-02-04)
+
+- fix: get feature name from /region_name or /product
+
+## 0.4.34 (2026-01-30)
+
+- fix: support '=' in LOCUS, COMMENT, etc
 
 ## 0.4.33 (2025-01-05)